Location_Aware_Model/backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Bcrypt==1.0.1
pymongo==4.6.1
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
xgboost==2.0.3
joblib==1.3.2
python-dotenv==1.0.0
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
reportlab==4.0.7
openpyxl==3.1.2
matplotlib==3.8.2
Pillow==10.2.0